lightnvm: fix up gc
The GC method for issuing IOs was broken after moving to be within
the block layer. Fix up the GC by letting GC IOs be submitted
through the queues ->make_request and let lightnvm know its a GC
by marking it with REQ_NVM_NO_INFLIGT.

The reason NO_INFLIGHT is used to disable inflight tracking, as it is
done by the GC, and thus an logical address is locked across multiple
IOs.