lightnvm: refactor base into core and target parts
To enable global garbage collection and multiple targets on a single
SSD, the implementation is split into two parts. A core part that is
integrated directly in to the block layer, and another having specific
targets.

The core part identifies the devices, and exposes a nvm_dev structure
for the targets to use. The targets may then retrieve this on
initialization.

This change now require targets to register themselves. So far there
isn't any automatic method, and thus /sys/block/dev/nvm/register must be
used to enable a specific target.

This is currently done by

  echo "rrpc test 0:0" > /sys/block/dev/nvm/register