lightnvm: introduce nvm_for_each_lun_ppa() macro
Users wish to iterate all luns and create a struct ppa_addr with
channel and lun set. To set the variables, two levels of loops are
required, one to iterate channels, and another to iterate luns. This
results in two-level deep for loops, which decreases readability.

Introduce nvm_for_each_lun_ppa, which implemets the nested loop and
sets ppa channel and lun variable for each loop body, eliminating
the necessary boilerplate code.