lightnvm: set block priority in nvm_set_rqd_ppalist()
When a ppa_list with vblks is expanded, it can either go block or
plane first. Currently it is been plane first. This works well for a
single vblk that is expanded, but fails for controllers that
require data for all planes is grouped together. Make the ppa_list
expand first expand with block, and then plane to support this
behavior.

 Before (2 blocks, 4 planes per vblk)
  idx block plane
   0      0     0
   1      1     0
   2      0     1
   3      1     1
   4      0     2
   5      1     2
   6      0     3
   7      1     3

 After
  idx block plane
   0      0     0
   1      0     1
   2      0     2
   3      0     3
   4      1     0
   5      1     1
   6      1     2
   7      1     3