lightnvm: set default lun range when no luns are specified
The create target ioctl takes a lun begin and lun end parameter, which
defines the range of luns to initialize a target with. If the user does
not define the parameters, it default to only using lun 0. Instead,
default to use the full range of the SSD, as it is the usual behavior
users want.