lightnvm: pblk: remove support for log lines
In the first version of pblk, we envisioned lines dedicated to store
metadata (log line). Instead, we move towards a model where distribute
the metadata at the beginning and end of all lines (used for user data).
This allows to optimize scheduling, minimizing collisions at the media
level.

This patch eliminates unused variables dedicated to track log lines and
renames what used to be data lines to improve code readability.

Signed-off-by: Javier González <javier@cnexlabs.com>