lightnvm: pblk: refactor locking for line lists
pblk uses a number of free and closed lists to organize lines. When the
current line becomes full, a new line is (i) taken from the free list,
(ii) prepared for usage and (iii) set to replace the current line for
future writes. Until now, a single lock was used for this process. Since
only one line is allocated at a time, this did represent a contention
point.

In preparation for introducing a wear-leveling index on pblk, replace
the locking strategy so that access to the free list is governed by a
dedicated lock. Indeed, since line preparation occurs in the write
thread - which is the user of the line -, locking is not necessary for this
part. Still, provide helper functions to lock this on exceptional cases
such as during recovery.

As part of the locking redesign, refactor the affected mapping
functions to avoid code replication.

Signed-off-by: Javier González <javier@cnexlabs.com>