lightnvm: pblk: support variable OOB size
pblk uses 8 bytes in the metadata region area exposed by the device
through the out of band area to store the lba mapped to the given
physical sector. This is used for recovery purposes. Given that the
first generation OCSSD devices exposed 16 bytes, pblk used a hard-coded
structure for this purpose.

This patch relaxes the 16 bytes assumption and uses the metadata size
reported by the device to layout metadata appropriately for the vector
commands. This adds support for arbitrary metadata sizes, as long as
these are larger than 8 bytes. Note that this patch does not address the
case in which the device does not expose an out of band area and that
pblk creation will fail in this case.

Signed-off-by: Javier González <javier@cnexlabs.com>